refactor: restore batch 4 ownership boundaries
