+++ packages/application/src/authority/index.ts
@@ -31,6 +31,7 @@ export * from "./canonical-cbor.ts";
 export * from "./semantic-mutation-envelope-v2.ts";
 export { authorityProtocolTuple } from "./types.ts";
 export * from "./cutover-control.ts";
+export * from "./daemon-host-contract.ts";
 export { createAuthorityCutoverEntityRegistryQualification } from "./cutover-registry-qualification.ts";
 export type {
   AttributedCoordinatorFactory,
