+++ packages/cli/src/daemon/command-service.ts
@@ -5,11 +5,12 @@ import {
   type AuthorityConnectionDispatch,
   type JsonObject
 } from "@harness-anything/daemon";
-import { makeHumanFallbackSessionProbe, makeTaskHolderService, type AuthorityCutoverControlService, type ProvenanceSessionExporterRejected, type ProvenanceSessionExportResult, type TaskHolderExecutor } from "@harness-anything/application";
+import { makeHumanFallbackSessionProbe, makeTaskHolderService, type AuthorityCutoverCommandAction, type AuthorityCutoverControlService, type ProvenanceSessionExporterRejected, type ProvenanceSessionExportResult, type TaskHolderExecutor } from "@harness-anything/application";
 import type { CurrentSessionRef, WriteCoordinator } from "@harness-anything/kernel";
 import { cliError, CliErrorCode } from "../cli/error-codes.ts";
 import { isDryRunAction } from "../cli/dry-run-preview.ts";
 import { normalizeCommandSemantics } from "../cli/command-semantic-normalizer.ts";
+import { productionAuthorityIngressFor } from "../cli/command-spec/index.ts";
 import { toCommandReceipt, type CommandFailureReceipt, type CommandReceipt } from "../cli/receipt.ts";
 import type { ParsedCommand } from "../cli/types.ts";
 import { isPlainRecord } from "../cli/value-utils.ts";
@@ -62,7 +63,7 @@ export function createCliCommandService(runtime: CliDaemonRuntime, options: CliC
         const daemonActor = context?.actor;
         if (isAuthorityCutoverAction(parsedCommand.action)) {
           return toCommandReceipt(await runAuthorityCutoverControlCommand({
-            action: parsedCommand.action,
+            action: parsedCommand.action as AuthorityCutoverCommandAction,
             control: options.authorityCutoverControl,
             authenticated: daemonActor !== undefined
           }));
@@ -83,7 +84,8 @@ export function createCliCommandService(runtime: CliDaemonRuntime, options: CliC
           ? makeDaemonAuthorityWriteCoordinator(authoritySubmissionV2, {
             command: parsedCommand,
             attribution,
-            currentSession
+            currentSession,
+            ingressAdapter: typedAuthorityIngressAdapter(parsedCommand.action.kind)
           })
           : undefined;
         const dryRun = isDryRunAction(parsedCommand.action);
@@ -152,6 +154,11 @@ export function createCliCommandService(runtime: CliDaemonRuntime, options: CliC
       }
     }
   };
+}
+
+function typedAuthorityIngressAdapter(kind: string) {
+  const ingress = productionAuthorityIngressFor(kind);
+  return ingress?.status === "typed-v2" ? ingress.adapter : undefined;
 }
 
 export function materializeExportedSession(
