+++ packages/daemon/src/authority/authority-cutover-command.ts
@@ -1,15 +1,18 @@
 // @slice-activation PLT-Boundary W2 daemon-owned authority cutover command host.
-import type { AuthorityCutoverControlService } from "@harness-anything/application";
-import type { AuthorityCutoverCommandErrorCode, AuthorityCutoverCommandResult, ParsedCommand } from "./command-contract.ts";
+import type {
+  AuthorityCutoverCommandAction,
+  AuthorityCutoverCommandErrorCode,
+  AuthorityCutoverCommandReport,
+  AuthorityCutoverCommandResult,
+  AuthorityCutoverControlService
+} from "@harness-anything/application";
 
-type AuthorityCutoverAction = Extract<ParsedCommand["action"], { readonly kind: `authority-cutover-${string}` }>;
-
-export function isAuthorityCutoverAction(action: ParsedCommand["action"]): action is AuthorityCutoverAction {
-  return action.kind.startsWith("authority-cutover-");
+export function isAuthorityCutoverAction(action: { readonly kind: string }): boolean {
+  return cutoverActionKinds.has(action.kind as AuthorityCutoverCommandAction["kind"]);
 }
 
 export async function runAuthorityCutoverControlCommand(input: {
-  readonly action: AuthorityCutoverAction;
+  readonly action: AuthorityCutoverCommandAction;
   readonly control?: AuthorityCutoverControlService;
   readonly authenticated: boolean;
 }): Promise<AuthorityCutoverCommandResult> {
@@ -18,18 +21,15 @@ export async function runAuthorityCutoverControlCommand(input: {
   try {
     const report = await executeAuthorityCutoverAction(input.control, input.action);
     const rejected = isRejectedCutoverReport(report);
-    return {
-      ok: !rejected,
-      command: input.action.kind,
-      report,
-      ...(rejected ? { error: { code: "write_rejected" as const, hint: cutoverRejectionMessage(report) } } : {})
-    };
+    return rejected
+      ? { ok: false, command: input.action.kind, report, error: { code: "write_rejected", hint: cutoverRejectionMessage(report) } }
+      : { ok: true, command: input.action.kind, report };
   } catch (error) {
     return cutoverCommandFailure(input.action.kind, "write_rejected", error instanceof Error ? error.message : String(error));
   }
 }
 
-async function executeAuthorityCutoverAction(control: AuthorityCutoverControlService, action: AuthorityCutoverAction): Promise<unknown> {
+async function executeAuthorityCutoverAction(control: AuthorityCutoverControlService, action: AuthorityCutoverCommandAction): Promise<AuthorityCutoverCommandReport> {
   if (action.kind === "authority-cutover-status") return control.status();
   if (action.kind === "authority-cutover-drain") return control.drain({ classifications: action.classifications });
   if (action.kind === "authority-cutover-scan") return control.scan({ profileId: action.profileId });
@@ -51,19 +51,27 @@ async function executeAuthorityCutoverAction(control: AuthorityCutoverControlSer
   });
 }
 
-function isRejectedCutoverReport(report: unknown): boolean {
-  if (!report || typeof report !== "object") return false;
-  const status = (report as { readonly status?: unknown }).status;
-  return status === "BLOCKED_UNCLASSIFIED_OPERATIONS" || status === "FINAL_SCAN_MISMATCH";
+function isRejectedCutoverReport(report: AuthorityCutoverCommandReport): boolean {
+  return (report.schema === "authority-cutover-drain-receipt/v1" && report.status === "BLOCKED_UNCLASSIFIED_OPERATIONS")
+    || (report.schema === "authority-cutover-equality-receipt/v1" && report.status === "FINAL_SCAN_MISMATCH");
 }
 
-function cutoverRejectionMessage(report: unknown): string {
-  const status = (report as { readonly status?: unknown }).status;
-  return status === "FINAL_SCAN_MISMATCH"
+function cutoverRejectionMessage(report: AuthorityCutoverCommandReport): string {
+  return report.schema === "authority-cutover-equality-receipt/v1"
     ? "Independent production final scans do not match; the cutover boundary remains closed."
     : "Authority drain has unclassified non-terminal operations; admission remains closed."
 }
 
 function cutoverCommandFailure(command: string, code: AuthorityCutoverCommandErrorCode, message: string): AuthorityCutoverCommandResult {
-  return { ok: false, command, error: { code, hint: message } };
+  return { ok: false, command: command as AuthorityCutoverCommandAction["kind"], error: { code, hint: message } };
 }
+
+const cutoverActionKinds = new Set<AuthorityCutoverCommandAction["kind"]>([
+  "authority-cutover-status",
+  "authority-cutover-drain",
+  "authority-cutover-scan",
+  "authority-cutover-confirm",
+  "authority-cutover-boundary",
+  "authority-cutover-freeze",
+  "authority-cutover-re-enable"
+]);
