+++ packages/daemon/src/lifecycle/compound-receipt-composition.ts
@@ -17,7 +17,6 @@ import {
   createBrokerCompoundReceiptCoordinatorV2,
   ReplicaBroker
 } from "@harness-anything/daemon";
-import { renderCompoundCliExit } from "@harness-anything/application";
 import { createDurableCompoundReceiptStoreV2 } from "./durable-compound-receipt-store.ts";
 
 /**
@@ -33,7 +32,6 @@ export interface ProductionCompoundReceiptComposition {
     readonly preparedReceiptDigest: string;
   }) => Promise<AckCommittedFrameV1>;
   readonly recover: (input: Omit<GetWaiterFrameV1, "type" | "kind">) => Promise<WaiterStateFrameV1>;
-  readonly renderExit: (receipt: CompoundOperationReceiptV2) => ReturnType<typeof renderCompoundCliExit>;
 }
 
 export function createProductionCompoundReceiptComposition(input: {
@@ -86,8 +84,7 @@ export function createProductionCompoundReceiptComposition(input: {
       const result = await coordinator.wire.handle({ type: "harness-compound-receipt-wire/v1", kind: "GET_WAITER", ...frame });
       if (result.kind !== "WAITER_STATE") throw new Error("COMPOUND_WAITER_QUERY_PROTOCOL_DAMAGED");
       return result;
-    },
-    renderExit: (receipt) => renderCompoundCliExit({ kind: "RECEIPT", receipt })
+    }
   };
 }
