+++ packages/daemon/test/fixtures/batch4-equivalence-golden.json
@@ -1,0 +1,9 @@
+{
+  "authorityWireErrors": "[\"AUTHORITY_PRODUCTION_LIFECYCLE_DISABLED\",\"AUTHORITY_WIRE_REPO_UNAVAILABLE\"]",
+  "daemonDrainTimeout": "{\"name\":\"DaemonDrainTimeoutError\",\"message\":\"daemon queue drain timed out after 1ms for daemon authority/runtime drain: unknown in-flight operation\"}",
+  "ownerExitReceipt": "{\"phase\":\"failed\",\"failure\":{\"code\":\"daemon_queue_drain_timeout\",\"hint\":\"Daemon restart requires the write queue to drain within the deadline, but in-flight operations failed to settle in time. Run `ha daemon status --json`, inspect the reported queue operation tuples, resolve or recover them, then retry the control request.\"}}",
+  "queuedCoordinatorError": "{\"_tag\":\"WriteRejected\",\"reason\":\"golden queued rejection\",\"code\":\"golden_queue_error\",\"retryable\":true}",
+  "compoundInternalError": "{\"exitCode\":1,\"symbol\":\"INTERNAL_ERROR\",\"stderr\":\"INTERNAL_ERROR: The effect or delivery outcome is not safely classified. Next: Query the opId and waiter first; resync when instructed.\",\"nextAction\":\"Query the opId and waiter first; resync when instructed.\",\"json\":{\"symbol\":\"INTERNAL_ERROR\",\"exitCode\":1,\"authority\":null,\"origin\":null,\"phase\":null,\"delivery\":null,\"historicalCut\":null,\"currentLease\":null,\"acknowledgement\":null,\"nextAction\":\"Query the opId and waiter first; resync when instructed.\"}}",
+  "compoundTerminalReceipt": "{\"exitCode\":20,\"symbol\":\"NOT_COMMITTED\",\"stderr\":\"NOT_COMMITTED: The authority terminally rejected the operation or found a base conflict. Next: Correct the request and submit with a new opId.\",\"nextAction\":\"Correct the request and submit with a new opId.\",\"json\":{\"symbol\":\"NOT_COMMITTED\",\"exitCode\":20,\"authority\":{\"tag\":\"REJECTED\",\"workspaceId\":\"workspace-golden\",\"opId\":\"op-golden\",\"semanticDigest\":\"aa\",\"reason\":\"golden rejection\"},\"origin\":null,\"phase\":\"COMMITTED\",\"delivery\":\"PENDING\",\"historicalCut\":null,\"currentLease\":\"NOT_REQUESTED\",\"acknowledgement\":null,\"nextAction\":\"Correct the request and submit with a new opId.\"}}",
+  "launchColdStartReceipt": "{\"status\":1,\"code\":\"AUTHORITY_MANIFEST_REGISTRY_INCOMPLETE\",\"missingOption\":\"--authority-manifest\",\"recovery\":\"ha daemon start --service --user-root <user-root> --authority-manifest <path>\"}"
+}
